wip: refactor extension
