+++ src/syrupy/__init__.py
@@ -7,8 +7,8 @@ from typing import (
 import pytest
 
 from .assertion import SnapshotAssertion
+from .extensions import DEFAULT_EXTENSION
 from .location import TestLocation
-from .serializers import DEFAULT_SERIALIZER
 from .session import SnapshotSession
 from .terminal import (
     green,
@@ -99,7 +99,7 @@ def pytest_sessionfinish(session: Any, exitstatus: int) -> None:
 def snapshot(request: Any) -> "SnapshotAssertion":
     return SnapshotAssertion(
         update_snapshots=request.config.option.update_snapshots,
-        serializer_class=DEFAULT_SERIALIZER,
+        serializer_class=DEFAULT_EXTENSION,
         test_location=TestLocation(request.node),
         session=request.session._syrupy,
     )
