+++ src/syrupy/extensions/__init__.py
@@ -1,4 +1,4 @@
 from .amber import AmberSnapshotSerializer
 
 
-DEFAULT_SERIALIZER = AmberSnapshotSerializer
+DEFAULT_EXTENSION = AmberSnapshotSerializer
