+++ src/syrupy/extensions/image.py
